pyproject.toml
[tool.poetry]
name = "hotel-assistant"
version = "0.1.0"
description = ""
authors = ["wei wei <cleaner.leon@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
langchain-openai = "^0.1.20"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.32"
langchain-core = "^0.2.36"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false
